HAAR-2226: renamed class
